requirements.txt
oauthlib==1.0.3
peewee==2.8.0
py-trello==0.4.3
python-dateutil==2.4.2
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
wheel==0.24.0
